Chỉnh sửa view và model

+++ WebSUDO_v2/Models/Entity/ChapterTruyen.cs
@@ -17,7 +17,7 @@ namespace Models.Entity
         [Key]
         [Column(Order = 1)]
         [DatabaseGenerated(DatabaseGeneratedOption.None)]
-        public int ChapterID { get; set; }
+        public long ChapterID { get; set; }
 
         [StringLength(250)]
         public string TenChapter { get; set; }
